[refactor] Handler 패키지 구조 변경 (factory, impl 패키지 생성)
