+++ shell/src/main/java/com/samsung/handler/impl/FileCommandHandler.java
@@ -1,6 +1,7 @@
-package com.samsung.handler;
+package com.samsung.handler.impl;
 
 import com.samsung.command.CommandInvoker;
+import com.samsung.handler.CommandHandler;
 
 import java.nio.file.Files;
 import java.nio.file.Paths;
