+++ shell/src/test/java/com/samsung/handler/factory/CommandHandlerFactoryTest.java
@@ -1,5 +1,8 @@
-package com.samsung.handler;
+package com.samsung.handler.factory;
 
+import com.samsung.handler.CommandHandler;
+import com.samsung.handler.impl.FileCommandHandler;
+import com.samsung.handler.impl.InteractiveCommandHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
